Parity with clojure.test
With these changes the ClojureScript test type seems pretty usable. So far only
tested with `cljs.repl.node` and `cljs.repl.browser`, I'm guessing those will be
the main ones people will be interested in.

- let ClojureScript render failure messages
- capture stack traces and send them to the reporter
- Fix browser compatibility (wait for scripts to be loaded)
- capture and forward "testing contexts" (nested `clojure.test/testing` calls)
- Support the "zero assertions" check
- Support the "single arity equals" check
- Prefix ids with "cljs", useful for CLJC
- Initiate the websocket connection directly from the client file, rather than
  through the REPL (prevents race conditions)

This change requires the latest Kaocha, as it relies on being able to send a
fully formatted failed assertion to the reporter, through the
`:kaocha.report/printed-expression` key.