Introduce the queue-eval-loop
We were jumping through quite some hoops to be able to use the ClojureScript
PREPL programmitcally (oh the irony) because it expects a Reader for its input,
which isn't that convenient when eval'ing forms that are coming from code,
rather than from some character stream.

Instead replaced the prepl with a "queue-eval-loop", which does a queue/take!
rather than a read+string. Seems to work well enough so far, let's hope it helps
with reliability.