Provide fallback file/line based on the deftest metadata
cljs.test has logic to detect the file/line number based on an exception's
stacktrace, this is used when an exception happens inside a deftest but outside
an `is` assertion. This logic is extremely flawed. Maybe it works on some JS
environments, but most of the time it just provides gibberish.

Since we know which test we are currently running we know at least the file/line
of the current deftest. So if it looks like :file/:line in the test event are
gibberish, then use the file/line of the deftest.